+++ src/core_modules/capture-core/components/ScopeSelector/ScopeSelector.container.js
@@ -1,112 +1,88 @@
 // @flow
-import React, { type ComponentType, useEffect, useCallback } from 'react';
+import React, { type ComponentType, useEffect, useCallback, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { useLocation } from 'react-router';
 import { ScopeSelectorComponent } from './ScopeSelector.component';
 import {
-    setOrgUnitFromLockedSelector,
-    setProgramIdFromLockedSelector,
-    setCategoryOptionFromLockedSelector,
-    resetCategoryOptionFromLockedSelector,
-    resetAllCategoryOptionsFromLockedSelector,
-    openNewRegistrationPageFromLockedSelector,
-    openSearchPageFromLockedSelector,
-    fetchOrgUnit,
+    setCategoryOptionFromScopeSelector,
+    resetCategoryOptionFromScopeSelector,
+    resetAllCategoryOptionsFromScopeSelector,
+    openNewRegistrationPageFromScopeSelector,
+    openSearchPageFromScopeSelector,
     resetProgramIdBatchAction,
-    startAgainBatchAction,
     resetOrgUnitIdBatchAction,
 } from './ScopeSelector.actions';
 import { resetProgramIdBase } from './QuickSelector/actions/QuickSelector.actions';
 import type { OwnProps } from './ScopeSelector.types';
-import { pageFetchesOrgUnitUsingTheOldWay } from '../../utils/url';
+import { useReset, useResetProgramId, useOrganizationUnit } from './hooks';
 
-const deriveReadiness = (lockedSelectorLoads, selectedOrgUnitId, organisationUnits) => {
+
+const deriveReadiness = (lockedSelectorLoads, selectedOrgUnitId, selectedOrgUnitName) => {
     // because we want the orgUnit to be fetched and stored
     // before allowing the user to view the locked selector
-    if (selectedOrgUnitId) {
-        const orgUnit = organisationUnits[selectedOrgUnitId];
-        return Boolean(orgUnit && orgUnit.id && !lockedSelectorLoads);
+    if (selectedOrgUnitId && selectedOrgUnitName) {
+        return true;
     }
     return !lockedSelectorLoads;
-};
-
-const useUrlQueries = (): { selectedProgramId: string, selectedOrgUnitId: string, pathname: string } =>
-    useSelector(({
-        currentSelections: {
-            programId: selectedProgramId,
-            orgUnitId: selectedOrgUnitId,
-        },
-        router: {
-            location: {
-                query: {
-                    programId: routerProgramId,
-                    orgUnitId: routerOrgUnitId,
-                },
-                pathname,
-            },
-        },
-    }) =>
-        ({
-            selectedProgramId: routerProgramId || selectedProgramId,
-            selectedOrgUnitId: routerOrgUnitId || selectedOrgUnitId,
-            pathname,
-        }),
-    );
-
-const useComponentLifecycle = () => {
-    const dispatch = useDispatch();
-    const { selectedOrgUnitId } = useUrlQueries();
-    const { pathname } = useLocation();
-    const pageFetchesOrgUnit = !pageFetchesOrgUnitUsingTheOldWay(pathname.substring(1));
-
-    useEffect(() => {
-        pageFetchesOrgUnit && selectedOrgUnitId && dispatch(fetchOrgUnit(selectedOrgUnitId));
-    },
-    [dispatch, selectedOrgUnitId, pageFetchesOrgUnit]);
 };
 
 export const ScopeSelector: ComponentType<OwnProps> =
   ({
       customActionsOnProgramIdReset = [],
       customActionsOnOrgUnitIdReset = [],
-      pageToPush = '',
       isUserInteractionInProgress = false,
+      selectedProgramId,
+      selectedOrgUnitId,
+      onSetProgramId,
+      onSetOrgUnit,
+      onResetProgramId,
+      onResetOrgUnitId,
+      children,
   }) => {
       const dispatch = useDispatch();
+      const { reset } = useReset();
+      const { resetProgramId } = useResetProgramId();
+      const { refetch: refetchOrganisationUnit, displayName } = useOrganizationUnit();
+      const [selectedOrgUnit, setSelectedOrgUnit] = useState({ name: displayName, id: selectedOrgUnitId });
 
-      const dispatchOnSetOrgUnit = useCallback(
-          (id: string, orgUnit: Object) => {
-              dispatch(setOrgUnitFromLockedSelector(id, orgUnit, pageToPush));
-          },
-          [pageToPush, dispatch]);
+      useEffect(() => {
+          const missName = !selectedOrgUnit.name;
+          const hasDifferentId = selectedOrgUnit.id !== selectedOrgUnitId;
 
-      const dispatchOnSetProgramId = useCallback(
-          (id: string) => {
-              dispatch(setProgramIdFromLockedSelector(id, pageToPush));
-          },
-          [pageToPush, dispatch]);
+          selectedOrgUnitId && (hasDifferentId || missName) && refetchOrganisationUnit({ variables: { selectedOrgUnitId } });
+      },
+      [selectedOrgUnitId]); // eslint-disable-line react-hooks/exhaustive-deps
+
+      useEffect(() => {
+          displayName && setSelectedOrgUnit(prevSelectedOrgUnit => ({ ...prevSelectedOrgUnit, name: displayName }));
+      },
+      [displayName, setSelectedOrgUnit]);
+
+      const handleSetOrgUnit = (orgUnitId, orgUnitObject) => {
+          setSelectedOrgUnit(orgUnitObject);
+          onSetOrgUnit(orgUnitId);
+      };
 
       const dispatchOnSetCategoryOption = useCallback(
           (categoryId: string, categoryOption: Object) => {
-              dispatch(setCategoryOptionFromLockedSelector(categoryId, categoryOption));
+              dispatch(setCategoryOptionFromScopeSelector(categoryId, categoryOption));
           },
           [dispatch]);
 
       const dispatchOnResetCategoryOption = useCallback(
           (categoryId: string) => {
-              dispatch(resetCategoryOptionFromLockedSelector(categoryId));
+              dispatch(resetCategoryOptionFromScopeSelector(categoryId));
           },
           [dispatch]);
 
       const dispatchOnResetAllCategoryOptions = useCallback(
           () => {
-              dispatch(resetAllCategoryOptionsFromLockedSelector());
+              dispatch(resetAllCategoryOptionsFromScopeSelector());
           },
           [dispatch]);
 
       const dispatchOnOpenNewEventPage = useCallback(
           () => {
-              dispatch(openNewRegistrationPageFromLockedSelector());
+              dispatch(openNewRegistrationPageFromScopeSelector());
           },
           [dispatch]);
 
@@ -114,15 +90,16 @@ export const ScopeSelector: ComponentType<OwnProps> =
           () => {
               const actions = [
                   resetProgramIdBase(),
-                  openNewRegistrationPageFromLockedSelector(),
+                  openNewRegistrationPageFromScopeSelector(),
               ];
-              dispatch(resetProgramIdBatchAction(actions, 'new'));
+              dispatch(resetProgramIdBatchAction(actions));
+              resetProgramId('new');
           },
-          [dispatch]);
+          [dispatch, resetProgramId]);
 
       const dispatchOnOpenSearchPage = useCallback(
           () => {
-              dispatch(openSearchPageFromLockedSelector());
+              dispatch(openSearchPageFromScopeSelector());
           },
           [dispatch]);
 
@@ -130,23 +107,26 @@ export const ScopeSelector: ComponentType<OwnProps> =
           () => {
               const actions = [
                   resetProgramIdBase(),
-                  openSearchPageFromLockedSelector(),
+                  openSearchPageFromScopeSelector(),
                   ...customActionsOnProgramIdReset,
               ];
-              dispatch(resetProgramIdBatchAction(actions, 'search'));
+              dispatch(resetProgramIdBatchAction(actions));
+              resetProgramId('search');
           },
-          [customActionsOnProgramIdReset, dispatch]);
+          [customActionsOnProgramIdReset, dispatch, resetProgramId]);
 
       const dispatchOnStartAgain = useCallback(
           () => {
-              dispatch(startAgainBatchAction());
-          }, [dispatch]);
+              dispatch(resetAllCategoryOptionsFromScopeSelector());
+              reset();
+          }, [dispatch, reset]);
 
       const dispatchOnResetOrgUnitId = useCallback(
           () => {
-              dispatch(resetOrgUnitIdBatchAction(customActionsOnOrgUnitIdReset, pageToPush));
+              dispatch(resetOrgUnitIdBatchAction(customActionsOnOrgUnitIdReset));
+              onResetOrgUnitId();
           },
-          [pageToPush, customActionsOnOrgUnitIdReset, dispatch]);
+          [customActionsOnOrgUnitIdReset, dispatch, onResetOrgUnitId]);
 
       const dispatchOnResetProgramId = useCallback(
           (baseAction: ReduxAction<any, any>) => {
@@ -155,22 +135,15 @@ export const ScopeSelector: ComponentType<OwnProps> =
                   ...customActionsOnProgramIdReset,
               ];
 
-              dispatch(resetProgramIdBatchAction(actions, pageToPush));
+              dispatch(resetProgramIdBatchAction(actions));
+              onResetProgramId();
           },
-          [pageToPush, customActionsOnProgramIdReset, dispatch]);
-
-      const { selectedOrgUnitId, selectedProgramId } = useUrlQueries();
+          [customActionsOnProgramIdReset, dispatch, onResetProgramId]);
 
       const lockedSelectorLoads: string =
         useSelector(({ activePage }) => activePage.lockedSelectorLoads);
 
-
-      const organisationUnits: Object =
-        useSelector(({ organisationUnits: orgUnits }) => orgUnits);
-
-      const ready = deriveReadiness(lockedSelectorLoads, selectedOrgUnitId, organisationUnits);
-
-      useComponentLifecycle();
+      const ready = deriveReadiness(lockedSelectorLoads, selectedOrgUnitId, selectedOrgUnit.name);
 
       return (
           <ScopeSelectorComponent
@@ -184,12 +157,15 @@ export const ScopeSelector: ComponentType<OwnProps> =
               onResetAllCategoryOptions={dispatchOnResetAllCategoryOptions}
               onResetCategoryOption={dispatchOnResetCategoryOption}
               onSetCategoryOption={dispatchOnSetCategoryOption}
-              onSetProgramId={dispatchOnSetProgramId}
-              onSetOrgUnit={dispatchOnSetOrgUnit}
+              onSetProgramId={onSetProgramId}
+              onSetOrgUnit={handleSetOrgUnit}
+              selectedOrgUnit={selectedOrgUnit}
               selectedOrgUnitId={selectedOrgUnitId}
               selectedProgramId={selectedProgramId}
               isUserInteractionInProgress={isUserInteractionInProgress}
               ready={ready}
-          />
+          >
+              {children}
+          </ScopeSelectorComponent>
       );
   };