+++ src/core_modules/capture-core/components/WidgetStagesAndEvents/WidgetStagesAndEvents.component.js
@@ -5,7 +5,7 @@ import { Widget } from '../Widget';
 import { Stages } from './Stages';
 import type { Props } from './stagesAndEvents.types';
 
-export const WidgetStagesAndEvents = ({ className, stages, events }: Props) => {
+export const WidgetStagesAndEvents = ({ className, stages, events, ...passOnProps }: Props) => {
     const [open, setOpenStatus] = useState(true);
     return (
         <div
@@ -21,6 +21,7 @@ export const WidgetStagesAndEvents = ({ className, stages, events }: Props) => {
                 <Stages
                     stages={stages}
                     events={events}
+                    {...passOnProps}
                 />
             </Widget>
         </div>