+++ src/core_modules/capture-core/components/WorkingLists/WorkingListsCommon/actions/workingListsCommon.actions.js
@@ -2,7 +2,7 @@
 import { actionCreator } from '../../../../actions/actions.utils';
 import type {
     SharingSettings,
-} from '../../WorkingLists';
+} from '../../WorkingListsBase';
 
 
 export const workingListsCommonActionTypes = {