+++ src/core_modules/capture-core/components/WorkingLists/WorkingListsCommon/hooks/useDataSource.js
@@ -4,7 +4,7 @@ import log from 'loglevel';
 import { errorCreator } from 'capture-core-utils';
 import { typeof dataElementTypes } from '../../../../metaData';
 import { convertClientToList } from '../../../../converters';
-import type { DataSource } from '../../WorkingLists';
+import type { DataSource } from '../../WorkingListsBase';
 
 export const useDataSource = (
     records?: { [string]: any },