+++ src/core_modules/capture-core/components/WorkingLists/WorkingListsCommon/hooks/useViewHasTemplateChanges/useViewHasTemplateChanges.types.js
@@ -1,5 +1,5 @@
 // @flow
-import type { FiltersData } from '../../../WorkingLists';
+import type { FiltersData } from '../../../WorkingListsBase';
 import type { InitialViewConfig } from '../../types';
 
 type Column = {|