+++ src/core_modules/capture-core/rules/converters/optionSetsConverter.js
@@ -1,0 +1,31 @@
+// @flow
+import type {
+    CachedOption,
+    CachedOptionSet,
+} from 'capture-core/storageControllers/cache.types';
+import type {
+    OptionSet,
+    OptionSets,
+} from 'capture-core-utils/rulesEngine/rulesEngine.types';
+
+function convertOptionSet(optionSet: CachedOptionSet): OptionSet {
+    const options = optionSet.options.map((option: CachedOption) => ({
+        id: option.id,
+        code: option.code,
+        displayName: option.displayName,
+    }));
+    return {
+        id: optionSet.id,
+        displayName: optionSet.displayName,
+        options,
+    };
+}
+
+export function convertOptionSetsToRulesEngineFormat(
+    cachedOptionSets: { [id: string]: CachedOptionSet }): OptionSets {
+    const optionSets = {};
+    Object.keys(cachedOptionSets).forEach((key: string) => {
+        optionSets[key] = convertOptionSet(cachedOptionSets[key]);
+    });
+    return optionSets;
+}